+++ test/expression_test_set.dart
@@ -644,7 +644,7 @@ class ExpressionTests extends TestSet {
   /// Tests differentiation of default functions.
   void defFuncDifferentiation() {
     Variable x = new Variable('x');
-    Number base = new Number(2);
+    Number two = new Number(2);
     var diff = [
                  // Expression,  deriveTo, output, outputSimplified
                  [new Exponential(x), 'x',
@@ -655,7 +655,7 @@ class ExpressionTests extends TestSet {
                                       '1.0 / x'],
                  // TODO Simplify can't cancel out terms yet, so the
                  //      simplified version is still a but ugly:
-                 [new Log(base, x),   'x',
+                 [new Log(two, x),    'x',
                                       '((((1.0 / x) * ln(2.0)) - (ln(x) * (0.0 / 2.0))) / (ln(2.0) * ln(2.0)))',
                                       '(((1.0 / x) * ln(2.0)) / (ln(2.0) * ln(2.0)))'],
                                       //'1.0 / (x * ln(2.0))'],
@@ -664,13 +664,16 @@ class ExpressionTests extends TestSet {
                  //[new Sqrt(x),        'x', '0.0', '0.0'],
                  //[new Root(2, x),     'x', '0.0', '0.0'],
 
-                 [new Sin(x),         'x', 'cos(x)',  'cos(x)'],
-                 [new Cos(x),         'x', '-sin(x)', '-sin(x)'],
+                 [new Sin(x),         'x', 'cos(x) * 1.0',  'cos(x)'],
+                 [new Cos(x),         'x', '-sin(x) * 1.0', '-sin(x)'],
 
                  // TODO Tan is internally handled as sin/cos:
                  //[new Tan(x),          'x', '0.0',    '0.0']
 
-                 [new Abs(x),          'x', 'sgn(x)', 'sgn(x)']
+                 [new Abs(x),         'x', 'sgn(x) * 1.0', 'sgn(x)'],
+                 [new Abs(two * x),   'x',
+                                      'sgn(2.0 * x) * (2.0 * 1.0 + 0.0 * x)',
+                                      'sgn(2.0 * x) * 2.0']
                 ];
 
     for (List exprCase in diff) {